Move flags to flaggable interface
